Rename components and follow the naming convention
PascalCase - UI components
camelCase - logical components
